combine and simplify evidence files
